app: Change the opacity param of _pixels to gfloat
All the internal math is float, and avoiding the conversion
from double is much faster when using SSE math.

+++ app/operations/gimpoperationantierasemode.h
@@ -55,7 +55,7 @@ gboolean gimp_operation_anti_erase_mode_process_pixels (gfloat              *in,
                                                         gfloat              *layer,
                                                         gfloat              *mask,
                                                         gfloat              *out,
-                                                        gdouble              opacity,
+                                                        gfloat               opacity,
                                                         glong                samples,
                                                         const GeglRectangle *roi,
                                                         gint                 level);
