+++ app/operations/gimpoperationmultiplymode.c
@@ -75,7 +75,7 @@ gimp_operation_multiply_mode_process (GeglOperation       *operation,
                                       const GeglRectangle *roi,
                                       gint                 level)
 {
-  gdouble opacity = GIMP_OPERATION_POINT_LAYER_MODE (operation)->opacity;
+  gfloat opacity = GIMP_OPERATION_POINT_LAYER_MODE (operation)->opacity;
 
   return gimp_operation_multiply_mode_process_pixels (in_buf, aux_buf, aux2_buf, out_buf, opacity, samples, roi, level);
 }
@@ -85,7 +85,7 @@ gimp_operation_multiply_mode_process_pixels (gfloat              *in,
                                              gfloat              *layer,
                                              gfloat              *mask,
                                              gfloat              *out,
-                                             gdouble              opacity,
+                                             gfloat               opacity,
                                              glong                samples,
                                              const GeglRectangle *roi,
                                              gint                 level)
