+++ app/operations/gimpoperationsaturationmode.h
@@ -54,7 +54,7 @@ gboolean gimp_operation_saturation_mode_process_pixels (gfloat              *in,
                                                         gfloat              *layer,
                                                         gfloat              *mask,
                                                         gfloat              *out,
-                                                        gdouble              opacity,
+                                                        gfloat               opacity,
                                                         glong                samples,
                                                         const GeglRectangle *roi,
                                                         gint                 level);
